Migrating classes into same test script, removed old test script
